Add test coverage for OpGraph so we don’t have to look at stories
